now no longer requires new /updater_scripts/ subfolder in game folder
the two files used to keep track of version for add-ons that don't explicitly record it, ArcDPS's d3d9 md5 and GW2Radial's version.txt, are now placed in their respective folders in the addons folder, which should already be in the Guild Wars 2 folder. This removes the requirement for an extra top-level directory in the game folder.