fixing bug where addons would not update
addon needed to be in enabledAddonsNames to be updated, but enabledAddonsNames is empty until entries are added at the *end* of the install process. Changed to check for the inverse instead, so it is now exclusive of disabled addons and inclusive of everything else rather than inclusive of enabled addons and exclusive of everything else.